[상점] 구매 가능한 목록과 판매한 목록이 적용되었습니다.

+++ nekoyume/Assets/_Scripts/Model/Agent.cs
@@ -7,12 +7,13 @@ namespace Nekoyume.Model
 {
     public static class Agent
     {
-        public static readonly ReactiveProperty<int> Gold = new ReactiveProperty<int>(0);
+        public static readonly ReactiveProperty<decimal> Gold = new ReactiveProperty<decimal>(0);
         
         private static readonly List<IDisposable> Disposables = new List<IDisposable>();
 
         static Agent()
         {
+            // FixMe. 골드 동기화 방법 수정 필요.
             RewardGold.RewardGoldMyselfSubject.ObserveOnMainThread().Subscribe(_ => Gold.Value = _).AddTo(Disposables);
         }
     }
